+++ packages/core/src/components/next/Dropdown/hooks/useDropdownFiltering.ts
@@ -1,10 +1,10 @@
 import { useState, useEffect } from "react";
 import { normalizeOptions } from "../utils/dropdownUtils";
-import { type BaseListItemData } from "../../../BaseListItem";
+import { type BaseItemData } from "../../../BaseItem";
 import { type DropdownListGroup } from "../components/DropdownBaseList/DropdownBaseList.types";
 import { type DropdownGroupOption } from "../Dropdown.types";
 
-function useDropdownFiltering<Item extends BaseListItemData>(
+function useDropdownFiltering<Item extends BaseItemData>(
   options: DropdownGroupOption<Item>,
   filterOption?: (option: Item, inputValue: string) => boolean,
   showSelectedOptions?: boolean,